% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/establish_survey_dir.R
\name{establish_survey_dir}
\alias{establish_survey_dir}
\title{Establish Survey Direction}
\usage{
establish_survey_dir(my_location)
}
\arguments{
\item{my_location}{Path on local system}
}
\value{
list object
}
\description{
Jim, tell me what it does
}
